status_function/requirements.txt
adal==1.2.7 ; python_version >= "3.7" and python_version < "4.0"
azure-common==1.1.27 ; python_version >= "3.7" and python_version < "4.0"
azure-core==1.18.0 ; python_version >= "3.7" and python_version < "4.0"
azure-functions==1.7.2 ; python_version >= "3.7" and python_version < "4.0"
azure-graphrbac==0.61.1 ; python_version >= "3.7" and python_version < "4.0"
azure-identity==1.6.1 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-authorization==2.0.0 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-core==1.3.0 ; python_version >= "3.7" and python_version < "4.0"
azure-mgmt-subscription==1.0.0 ; python_version >= "3.7" and python_version < "4.0"
cachetools==5.2.1 ; python_version >= "3.7" and python_version < "4.0"
certifi==2021.5.30 ; python_version >= "3.7" and python_version < "4.0"
cffi==1.14.6 ; python_version >= "3.7" and python_version < "4.0"
charset-normalizer==2.0.6 ; python_version >= "3.7" and python_version < "4.0"
cryptography==3.4.8 ; python_version >= "3.7" and python_version < "4.0"
google-api-core==2.8.2 ; python_version >= "3.7" and python_version < "4.0"
google-auth==2.16.0 ; python_version >= "3.7" and python_version < "4.0"
googleapis-common-protos==1.56.4 ; python_version >= "3.7" and python_version < "4.0"
idna==3.2 ; python_version >= "3.7" and python_version < "4.0"
isodate==0.6.0 ; python_version >= "3.7" and python_version < "4.0"
msal-extensions==0.3.0 ; python_version >= "3.7" and python_version < "4.0"
msal==1.14.0 ; python_version >= "3.7" and python_version < "4.0"
msrest==0.6.21 ; python_version >= "3.7" and python_version < "4.0"
msrestazure==0.6.4 ; python_version >= "3.7" and python_version < "4.0"
oauthlib==3.1.1 ; python_version >= "3.7" and python_version < "4.0"
opencensus-context==0.1.3 ; python_version >= "3.7" and python_version < "4.0"
opencensus-ext-azure==1.1.8 ; python_version >= "3.7" and python_version < "4.0"
opencensus==0.11.1 ; python_version >= "3.7" and python_version < "4.0"
portalocker==1.7.1 ; python_version >= "3.7" and python_version < "4.0"
protobuf==3.19.3 ; python_version >= "3.7" and python_version < "4.0"
psutil==5.9.4 ; python_version >= "3.7" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.7" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.7" and python_version < "4.0"
pycparser==2.20 ; python_version >= "3.7" and python_version < "4.0"
pydantic[dotenv]==1.8.2 ; python_version >= "3.7" and python_version < "4.0"
pyjwt==2.1.0 ; python_version >= "3.7" and python_version < "4.0"
pyjwt[crypto]==2.1.0 ; python_version >= "3.7" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.7" and python_version < "4.0"
python-dotenv==0.19.0 ; python_version >= "3.7" and python_version < "4.0"
requests-oauthlib==1.3.0 ; python_version >= "3.7" and python_version < "4.0"
requests==2.26.0 ; python_version >= "3.7" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.7" and python_version < "4"
six==1.16.0 ; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.2 ; python_version >= "3.7" and python_version < "4.0"
urllib3==1.26.7 ; python_version >= "3.7" and python_version < "4"
